inlined sources.dart and xiph.dart into build.dart
